* adapted test case due to API change in Hibernate 3.5 * references #221
git-svn-id: svn+ssh://svn.synyx.de/var/svn/synyx/opensource/hades/trunk@8844 5a64d73e-33d6-4ccc-9058-23f8668ecac9

+++ hades/src/main/java/org/synyx/hades/dao/orm/GenericHibernateJpaDao.java
@@ -113,7 +113,7 @@ public class GenericHibernateJpaDao<T extends Persistable<PK>, PK extends Serial
 
         Criteria countCriteria = applyExamples(examples);
         countCriteria.setProjection(Projections.rowCount());
-        Integer count = (Integer) countCriteria.uniqueResult();
+        Long count = (Long) countCriteria.uniqueResult();
 
         Criteria listCriteria = applyExamples(examples);
